Rename test folders & split pipeline repo unit tests into sep files
