+++ tests/non_search/routers/geographies/setup_world_map_helpers.py
