+++ tests/non_search/routers/lookups/test_geo_stats.py
