+++ tests/non_search/routers/test_auth.py
