+++ tests/non_search/routers/test_pipeline_trigger.py
