+++ tests/unit/app/core/test_aws.py
