+++ tests/unit/app/core/test_lookups.py
